Small refactoring and bug fixes on Leaderboards
